Bug 1567036 - Clear BidiParagraphData::mPrevFrame when ResolveParagraphWithinBlock creates a new continuation, which invalidates its relationship to the line iterator.  r?jfkthame
This patch adds an assertion that documents the invariant that this code
needs to maintain to manage the line iterator correctly.  Failing to
meet that invariant would also cause additional assertions in my
work-in-progress on bug 1300293.

It then adds two assignments of mPrevFrame to null that make that
assertion hold.

The first assignment to null is tested by a number of tests already in
our test suite, including
layout/reftests/w3c-css/received/css-writing-modes/block-plaintext-004.html
and layout/reftests/bidi/unicode-bidi-plaintext.html , which would hit
the added assertion if that assignment to null were not present.

The second assignment to null was untested in our test suite (though it
showed up when doing view-source: of https://html.spec.whatwg.org/ ), so
I've added a reftest (1567036-1.html) that hits the added assertion if
that second assignment to null is not present.

Differential Revision: https://phabricator.services.mozilla.com/D38438