Bug 1420528 - Add reftest.  r=TYLin
Note that I introduced a blank line to make the intent of the NOTE
clearer, which I had to research.  I think it's clear it covers the
three tests below it based on
https://hg.mozilla.org/mozilla-central/rev/53489b3e14f1 and
https://bugzilla.mozilla.org/show_bug.cgi?id=967311#c0 .

I'm also running this test both with and without the column-span pref,
because those two states will be fixed by different patches.

Co-authored-by: L. David Baron <dbaron@dbaron.org>
Co-authored-by: Daniel Holbert <dholbert@cs.stanford.edu>

Differential Revision: https://phabricator.services.mozilla.com/D39581