Bug 1547759 - Remove BCTableCell as a distinct frame type.  r?heycam
There are two existing callers of IsTableCellFrame that both appear to
want to include BCTableCell frames, but don't.  The patch three patches
ahead of this one will add a third that wants the same.

The existing users of frame types all have to work around it being a
distinct type, and none appear to want the distinction.  This removes
that complexity.

If any callers want to make the distinction, they could implement
QueryFrame for BCTableCellFrame and use it.

In a little more detail, prior to this patch:
  * nsQueryFrame::ClassID::nsMathMLmtdFrame_id was 46
  * nsQueryFrame::ClassID::nsTableCellFrame_id was 47
  * nsQueryFrame::ClassID::nsBCTableCellFrame_id was 65
and entries 46 and 47 in sLayoutFrameTypes were
mozilla::LayoutFrameType::TableCell while entry 65 was
mozilla::LayoutFrameType::BCTableCell.

With this patch:
  * nsQueryFrame::ClassID::nsBCTableCellFrame_id is 40
  * nsQueryFrame::ClassID::nsMathMLmtdFrame_id is 41
  * nsQueryFrame::ClassID::nsTableCellFrame_id is 42
and entries 40 through 42 in sLayoutFrameTypes are
mozilla::LayoutFrameType::TableCell.